+++ packages/axelar-soroban-std/src/interfaces/testdata/contract.rs
@@ -61,3 +61,32 @@ pub enum DataKey {
 pub enum ContractError {
     SomeFailure = 1,
 }
+
+mod test {
+    use soroban_sdk::{contracttype, Env};
+
+    use crate::interfaces::testdata::contract::DataKey;
+
+    use super::Contract;
+
+    #[test]
+    fn contracttype_enum_name_is_irrelevant_for_key_collision() {
+        let env = Env::default();
+        let contract_id = env.register(Contract, ());
+
+        env.as_contract(&contract_id, || {
+            assert!(!env.storage().instance().has(&DataKey::Migrating));
+            assert!(!env.storage().instance().has(&DataKey2::Migrating));
+
+            env.storage().instance().set(&DataKey::Migrating, &());
+
+            assert!(env.storage().instance().has(&DataKey::Migrating));
+            assert!(env.storage().instance().has(&DataKey2::Migrating));
+        });
+    }
+
+    #[contracttype]
+    enum DataKey2 {
+        Migrating,
+    }
+}
