+++ packages/axelar-soroban-std/src/interfaces/testdata/mod.rs
@@ -1,1 +1,3 @@
-pub mod contract;
+mod contract;
+
+pub use contract::{Contract, ContractClient};
