+++ packages/axelar-soroban-std/src/interfaces/upgradable.rs
@@ -147,41 +147,17 @@ mod test {
     use crate::interfaces::upgradable::{OwnershipClient, UpgradableClient, UpgradedEvent};
     use crate::{assert_invoke_auth_err, assert_invoke_auth_ok, events};
 
-    use crate::interfaces::testdata::contract::ContractClient;
+    use crate::interfaces::testdata::ContractClient;
     use crate::interfaces::{testdata, upgradable};
     use soroban_sdk::testutils::{Address as _, MockAuth, MockAuthInvoke};
-    use soroban_sdk::{contracttype, Address, Env, String};
+    use soroban_sdk::{Address, Env, String};
 
     const WASM: &[u8] = include_bytes!("testdata/contract.wasm");
 
     #[test]
-    fn contracttype_enum_name_is_irrelevant_for_key_collision() {
-        let env = Env::default();
-        let contract_id = env.register(testdata::contract::Contract, ());
-
-        env.as_contract(&contract_id, || {
-            assert!(!env
-                .storage()
-                .instance()
-                .has(&testdata::contract::DataKey::Migrating));
-            assert!(!env.storage().instance().has(&DataKey2::Migrating));
-
-            env.storage()
-                .instance()
-                .set(&testdata::contract::DataKey::Migrating, &());
-
-            assert!(env
-                .storage()
-                .instance()
-                .has(&testdata::contract::DataKey::Migrating));
-            assert!(env.storage().instance().has(&DataKey2::Migrating));
-        });
-    }
-
-    #[test]
     fn owner_fails_if_owner_not_set() {
         let env = Env::default();
-        let contract_id = env.register(testdata::contract::Contract, ());
+        let contract_id = env.register(testdata::Contract, ());
 
         assert!(OwnershipClient::new(&env, &contract_id)
             .try_owner()
@@ -191,7 +167,7 @@ mod test {
     #[test]
     fn owner_returns_correct_owner_when_set() {
         let env = Env::default();
-        let contract_id = env.register(testdata::contract::Contract, ());
+        let contract_id = env.register(testdata::Contract, ());
 
         let owner = Address::generate(&env);
         env.as_contract(&contract_id, || {
@@ -204,7 +180,7 @@ mod test {
     #[test]
     fn upgrade_fails_if_owner_not_set() {
         let env = Env::default();
-        let contract_id = env.register(testdata::contract::Contract, ());
+        let contract_id = env.register(testdata::Contract, ());
         let hash = env.deployer().upload_contract_wasm(WASM);
 
         assert!(UpgradableClient::new(&env, &contract_id)
@@ -215,7 +191,7 @@ mod test {
     #[test]
     fn upgrade_fails_if_caller_not_authenticated() {
         let env = Env::default();
-        let contract_id = env.register(testdata::contract::Contract, ());
+        let contract_id = env.register(testdata::Contract, ());
         let hash = env.deployer().upload_contract_wasm(WASM);
 
         let owner = Address::generate(&env);
@@ -231,7 +207,7 @@ mod test {
     #[test]
     fn upgrade_fails_if_called_by_non_owner() {
         let env = Env::default();
-        let contract_id = env.register(testdata::contract::Contract, ());
+        let contract_id = env.register(testdata::Contract, ());
         let hash = env.deployer().upload_contract_wasm(WASM);
 
         let owner = Address::generate(&env);
@@ -246,7 +222,7 @@ mod test {
     #[test]
     fn upgrade_succeeds_if_owner_is_authenticated() {
         let env = Env::default();
-        let contract_id = env.register(testdata::contract::Contract, ());
+        let contract_id = env.register(testdata::Contract, ());
         let hash = env.deployer().upload_contract_wasm(WASM);
 
         let owner = Address::generate(&env);
@@ -261,7 +237,7 @@ mod test {
     #[test]
     fn migrate_fails_if_caller_not_authenticated() {
         let env = Env::default();
-        let contract_id = env.register(testdata::contract::Contract, ());
+        let contract_id = env.register(testdata::Contract, ());
         let hash = env.deployer().upload_contract_wasm(WASM);
 
         let owner = Address::generate(&env);
@@ -279,7 +255,7 @@ mod test {
     #[test]
     fn migrate_fails_if_called_by_non_owner() {
         let env = Env::default();
-        let contract_id = env.register(testdata::contract::Contract, ());
+        let contract_id = env.register(testdata::Contract, ());
         let hash = env.deployer().upload_contract_wasm(WASM);
 
         let owner = Address::generate(&env);
@@ -297,7 +273,7 @@ mod test {
     #[test]
     fn migrate_fails_if_not_called_after_upgrade() {
         let env = Env::default();
-        let contract_id = env.register(testdata::contract::Contract, ());
+        let contract_id = env.register(testdata::Contract, ());
 
         let owner = Address::generate(&env);
         env.as_contract(&contract_id, || {
@@ -311,7 +287,7 @@ mod test {
     #[test]
     fn migrate_succeeds_if_owner_is_authenticated_and_called_after_upgrade() {
         let env = Env::default();
-        let contract_id = env.register(testdata::contract::Contract, ());
+        let contract_id = env.register(testdata::Contract, ());
         let hash = env.deployer().upload_contract_wasm(WASM);
 
         let owner = Address::generate(&env);
@@ -341,7 +317,7 @@ mod test {
     #[test]
     fn simulate_migration_for_code_coverage() {
         let env = Env::default();
-        let contract_id = env.register(testdata::contract::Contract, ());
+        let contract_id = env.register(testdata::Contract, ());
 
         let owner = Address::generate(&env);
         env.as_contract(&contract_id, || {
@@ -356,7 +332,7 @@ mod test {
     #[test]
     fn migrate_fails_if_called_twice() {
         let env = Env::default();
-        let contract_id = env.register(testdata::contract::Contract, ());
+        let contract_id = env.register(testdata::Contract, ());
         let hash = env.deployer().upload_contract_wasm(WASM);
 
         let owner = Address::generate(&env);
@@ -371,10 +347,5 @@ mod test {
         assert_invoke_auth_ok!(owner, client.try_migrate(&()));
 
         assert_invoke_auth_err!(owner, client.try_migrate(&()));
-    }
-
-    #[contracttype]
-    enum DataKey2 {
-        Migrating,
     }
 }
